// src/shared/interface.ts
export type RuneUTXO = {
  txId: string
  txIndex: string
  amountOfRunes: number
  address: string
  script: string
  satoshis: number
}